Rename integresql source files to PascalCase
Rename getConnection.ts, templateIdFromFilesHash.ts, globalSetup.ts,
and startContainers.ts to PascalCase to match effect-ts convention.
Update all imports and config references accordingly.
