+++ packages/integresql/src/examples.node.integration.test.ts
@@ -6,7 +6,7 @@ import { inject } from "vitest"
 import type { DatabaseConfiguration } from "./index.js"
 import { getConnection } from "./index.js"
 import { Sandbox } from "./Sandbox.ts"
-import { templateIdFromFiles } from "./templateIdFromFilesHash.js"
+import { templateIdFromFiles } from "./TemplateIdFromFilesHash.js"
 
 const dependencies = NodeContext.layer
