+++ packages/integresql/tsconfig.src.json
@@ -11,9 +11,9 @@
     "src"
   ],
   "exclude": [
-    "src/globalSetup.ts", 
+    "src/GlobalSetup.ts", 
     "src/Sandbox.ts",
-    "src/startContainers.ts",
+    "src/StartContainers.ts",
     "src/**/*.node.unit.test.ts",
     "src/**/*.node.integration.test.ts",
     "src/**/*.test-d.ts"
